move things to lib
